+++ examples/android/LuminaDemo/app/src/main/res/drawable/lumina_logo.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" xmlns:aapt="http://schemas.android.com/aapt"
+    android:viewportWidth="36.3"
+    android:viewportHeight="36"
+    android:width="36.3dp"
+    android:height="36dp">
+    <path
+        android:pathData="M23.3814 24c0.0009 0 0.0018 0 0.0027 0h0.2118c0.3231 0 0.585 0.26 0.585 0.5806v0.2103c0 0.0009 0 0.0018 0 0.0026V24.5806C24.1809 24.26 23.919 24 23.5959 24Z"
+        android:fillColor="#FFFFFF" />
+    <path
+        android:pathData="M12.0905 24.5806C12.0905 24.26 12.3524 24 12.6755 24H12.89 12.8873 12.6755c-0.3231 0 -0.585 0.26 -0.585 0.5806v0.2103 0.0026z"
+        android:fillColor="#FFFFFF" />
+    <path
+        android:pathData="M0 18.0827C0 13.2189 1.94367 8.8056099 5.10215 5.5663199c0.3517 -0.3607 0.94309 -0.10327 0.94309 0.39868V23.2892c0 0.4382 0.35796 0.7935 0.79953 0.7935h4.33893c0.3022 0.025 0.9068 0.24 0.9068 0.9v4.3064c0 0.4383 0.3579 0.7936 0.7995 0.7936h10.4914c0.4416 0 0.7995 -0.3553 0.7995 -0.7936v-4.3064c0 -0.66 0.6046 -0.875 0.9068 -0.9h4.3389c0.4416 0 0.7996 -0.3553 0.7996 -0.7935v-4.413c0 -0.4382 -0.358 -0.7935 -0.7996 -0.7935h-4.4461c-0.4416 0 -0.7996 0.3553 -0.7996 0.7935v4.3069c-0.0002 0.6597 -0.6046 0.8746 -0.9067 0.8996H12.9973c-0.3022 -0.025 -0.9066 -0.2399 -0.9068 -0.8996V1.8439899c0 -0.44879 0.2244 -0.82379 0.8501 -1.01128999 0.008 -0.002394 0.0944 -0.028308 0.0944 -0.028308 1.618 -0.469747 3.3297 -0.721692 5.1007 -0.721692 10.0161 0 18.1357 8.05887999 18.1357 18.00000009 0 9.9411 -8.1196 18 -18.1357 18C8.11963 36.0827 0 28.0238 0 18.0827Z"
+        android:fillColor="#1A1A1A" />
+</vector>
